Move hcl attribute implementation to new file
